# api/openapi.yaml
openapi: 3.1.0
info:
  title: Accident Reconstructor API
  version: 0.1.0
  description: Runtime API for vehicle accident reconstruction jobs and artifacts.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /readyz:
    get:
      operationId: getReady
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /api/v1/tools:
    get:
      operationId: listTools
      responses:
        "200":
          description: Toolchain status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolStatus"
  /api/v1/cases:
    get:
      operationId: listCases
      responses:
        "200":
          description: Cases held by this backend process
          content:
            application/json:
              schema:
                type: object
                required: [cases]
                properties:
                  cases:
                    type: array
                    items:
                      $ref: "#/components/schemas/CaseSummary"
    post:
      operationId: createCase
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [videos]
              properties:
                case_name:
                  type: string
                scale_meters:
                  type: number
                  format: double
                videos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "202":
          description: Case accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCaseResponse"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/cases/{caseId}:
    get:
      operationId: getCase
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Case summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseSummary"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/cases/{caseId}/artifact:
    get:
      operationId: getArtifact
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Reconstruction artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/cases/{caseId}/report:
    get:
      operationId: getReport
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Markdown report
          content:
            text/markdown:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/cases/{caseId}/bundle:
    get:
      operationId: getBundle
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: ZIP evidence bundle
          content:
            application/zip:
              schema:
                type: string
                format: binary
components:
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
    ToolStatus:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        path:
          type: string
        version:
          type: string
        status:
          type: string
    UploadInfo:
      type: object
      required:
        [fileName, sizeBytes, durationSeconds, width, height, frameRate, sha256]
      properties:
        fileName:
          type: string
        sizeBytes:
          type: integer
          format: int64
        durationSeconds:
          type: number
        width:
          type: integer
        height:
          type: integer
        frameRate:
          type: number
        sha256:
          type: string
    CaseSummary:
      type: object
      required:
        [id, name, status, message, progress, createdAt, updatedAt, uploads]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [queued, processing, complete, failed]
        message:
          type: string
        progress:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        uploads:
          type: array
          items:
            $ref: "#/components/schemas/UploadInfo"
        artifactUrl:
          type: string
        reportUrl:
          type: string
        error:
          $ref: "#/components/schemas/ErrorResponse"
    CreateCaseResponse:
      type: object
      required: [case]
      properties:
        case:
          $ref: "#/components/schemas/CaseSummary"
    Point3D:
      type: object
      required: [id, x, y, z, color]
      properties:
        id:
          type: string
        x:
          type: number
        y:
          type: number
        z:
          type: number
        color:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: integer
        tags:
          type: array
          items:
            type: string
    CameraPose:
      type: object
      required:
        [id, sourceVideo, timeSeconds, position, rotationEuler, focalLengthPx]
      properties:
        id:
          type: string
        sourceVideo:
          type: string
        timeSeconds:
          type: number
        position:
          $ref: "#/components/schemas/Vec3"
        rotationEuler:
          $ref: "#/components/schemas/Vec3"
        focalLengthPx:
          type: number
    VehicleTrackPoint:
      type: object
      required: [timeSeconds, x, y, z, speedMps]
      properties:
        timeSeconds:
          type: number
        x:
          type: number
        y:
          type: number
        z:
          type: number
        speedMps:
          type: number
    SpeedEstimate:
      type: object
      required:
        [
          method,
          meanMps,
          meanKph,
          meanMph,
          lowerKph,
          upperKph,
          confidence,
          notes,
        ]
      properties:
        method:
          type: string
        meanMps:
          type: number
        meanKph:
          type: number
        meanMph:
          type: number
        lowerKph:
          type: number
        upperKph:
          type: number
        confidence:
          type: number
        notes:
          type: array
          items:
            type: string
    ReconstructionQuality:
      type: object
      required: [mode, inputVideos, coordinateFrame, toolchain, warnings]
      properties:
        mode:
          type: string
        inputVideos:
          type: integer
        coordinateFrame:
          type: string
        toolchain:
          type: array
          items:
            $ref: "#/components/schemas/ToolStatus"
        warnings:
          type: array
          items:
            type: string
        reprojectionRmse:
          type: number
    Artifact:
      type: object
      required:
        - caseId
        - caseName
        - version
        - commit
        - createdAt
        - coordinateFrame
        - uploads
        - points
        - cameras
        - vehicleTrack
        - speed
        - quality
        - reportMarkdown
      properties:
        caseId:
          type: string
        caseName:
          type: string
        version:
          type: string
        commit:
          type: string
        createdAt:
          type: string
          format: date-time
        coordinateFrame:
          type: string
        uploads:
          type: array
          items:
            $ref: "#/components/schemas/UploadInfo"
        points:
          type: array
          items:
            $ref: "#/components/schemas/Point3D"
        cameras:
          type: array
          items:
            $ref: "#/components/schemas/CameraPose"
        vehicleTrack:
          type: array
          items:
            $ref: "#/components/schemas/VehicleTrackPoint"
        speed:
          $ref: "#/components/schemas/SpeedEstimate"
        quality:
          $ref: "#/components/schemas/ReconstructionQuality"
        reportMarkdown:
          type: string
    Vec3:
      type: array
      minItems: 3
      maxItems: 3
      items:
        type: number